+++ apps/backend/src/libs/exceptions/exceptions.ts
@@ -1,2 +1,3 @@
 export { MailError } from "./mail-error.exception.js";
+export { JWTExpired } from "jose/errors";
 export { ValidationError } from "shared";