+++ apps/backend/src/libs/modules/token/token.ts
@@ -5,7 +5,6 @@ import { BaseToken } from "./base-token.module.js";
 
 const token = new BaseToken<TokenPayload>({
 	algorithm: config.ENV.JWT.ALGORITHM,
-	expirationTime: config.ENV.JWT.EXPIRATION_TIME,
 	secret: Buffer.from(config.ENV.JWT.SECRET),
 });