+++ apps/frontend/src/libs/components/balance-wheel-chart/libs/enums/enums.ts
@@ -1,2 +1,13 @@
+export { AngleCoefficient } from "./angle-coefficient.enum.js";
 export { BalanceWheelChartDataConfig } from "./balance-wheel-data-config.enum.js";
+export { ChartFont } from "./chart-font.enum.js";
+export { ChartGraphicsColor } from "./chart-graphics-color.enum.js";
 export { ChartSliceColor } from "./chart-slice-color.enum.js";
+export { DotOffsetX } from "./dot-offset-x.enum.js";
+export { DotOffsetY } from "./dot-offset-y.enum.js";
+export { LabelDistacneOffset } from "./label-distance-offset.enum.js";
+export { SublabelOffsetX } from "./sublabel-offset-x.enum.js";
+export { SublabelOffsetY } from "./sublabel-offset-y.enum.js";
+export { ValueLabelOffsetX } from "./value-label-offset-x.enum.js";
+export { ValueLabelOffsetY } from "./value-label-offset-y.enum.js";
+export { WheelCenterDistance } from "./wheel-center-distance.enum.js";