+++ apps/frontend/src/libs/components/balance-wheel-chart/libs/helpers/helpers.ts
@@ -1,2 +1,3 @@
+export { drawExtraPointGraphics } from "./draw-extra-point-graphics/draw-extra-point-graphics.helper.js";
 export { generateGradientColor } from "./generate-gradient-color/generate-gradient-color.helper.js";
 export { generateRandomData } from "./generate-random-data/generate-random-data.helper.js";