+++ apps/frontend/src/pages/onboarding/libs/constants/constants.ts
@@ -1,1 +1,2 @@
+export { FIRST_QUESTION_INDEX } from "./first-question-index.constant.js";
 export { ONBOARDING_FORM_DEFAULT_VALUES } from "./onboarding-form-default-values.contant.js";