+++ apps/mobile/src/libs/components/planet/libs/maps/maps.ts
@@ -1,2 +1,1 @@
-export { directionToGradient } from "./direction-to-gradient";
 export { sizeToStyles } from "./size-to-styles";