+++ apps/mobile/src/packages/quiz/quiz-api.ts
@@ -1,0 +1,26 @@
+import { APIPath, ContentType } from "~/libs/enums/enums";
+import { type APIConfiguration, BaseHttpApi } from "~/libs/packages/api/api";
+
+import { QuizApiPath } from "./libs/enums/enums";
+import { type QuizQuestionDto } from "./libs/types/types";
+
+class QuizApi extends BaseHttpApi {
+	public constructor({ baseUrl, http, storage }: APIConfiguration) {
+		super({ baseUrl, http, path: APIPath.QUIZ, storage });
+	}
+
+	public async getAllQuestions(): Promise<{ items: QuizQuestionDto[][] }> {
+		const response = await this.load(
+			this.getFullEndpoint(QuizApiPath.QUESTIONS, {}),
+			{
+				contentType: ContentType.JSON,
+				hasAuth: true,
+				method: "GET",
+			},
+		);
+
+		return await response.json<{ items: QuizQuestionDto[][] }>();
+	}
+}
+
+export { QuizApi };