+++ apps/mobile/src/screens/onboarding/libs/components/content/styles.ts
@@ -1,0 +1,10 @@
+import { StyleSheet } from "react-native";
+
+const styles = StyleSheet.create({
+	title: {
+		marginVertical: 70,
+		textAlign: "center",
+	},
+});
+
+export { styles };