+++ apps/mobile/src/screens/quiz/libs/constants/costants.ts
@@ -1,0 +1,1 @@
+export { QUIZ_FORM_DEFAULT_VALUES } from "./quiz-form-default-values.constant";