+++ apps/mobile/src/screens/tasks/tasks.tsx
@@ -1,15 +1,28 @@
 import React from "react";
 
-import { ScreenWrapper, Tag, Text } from "~/libs/components/components";
-import { useAppDispatch, useEffect } from "~/libs/hooks/hooks";
+import {
+	ScreenWrapper,
+	SliderContent,
+	Tag,
+	Text,
+} from "~/libs/components/components";
+import { useAppDispatch, useAppSelector, useEffect } from "~/libs/hooks/hooks";
+import { type SliderData } from "~/libs/types/types";
+import { type UserDto } from "~/packages/users/users";
 import { actions as userActions } from "~/slices/users/users";
+
+const sliderData: SliderData[] = [{ color: "red", label: "Love" }];
 
 const Tasks: React.FC = () => {
 	const dispatch = useAppDispatch();
 
+	const authenticatedUser = useAppSelector(({ auth }) => auth.user);
+
 	useEffect(() => {
-		void dispatch(userActions.loadAll());
-	}, [dispatch]);
+		void dispatch(
+			userActions.getById({ id: (authenticatedUser as UserDto).id }),
+		);
+	}, [dispatch, authenticatedUser]);
 
 	return (
 		<ScreenWrapper>
@@ -23,6 +36,10 @@ const Tasks: React.FC = () => {
 			<Tag color="pink" label="Free time" />
 			<Tag color="orange" label="Spiritual" />
 			<Tag color="blue" label="Mental" />
+
+			{sliderData.map((item) => (
+				<SliderContent color={item.color} key={item.label} label={item.label} />
+			))}
 		</ScreenWrapper>
 	);
 };